add support for throwing warning for takes.

+++ source/backend/api/Services/AcquisitionFileService.cs
@@ -178,9 +178,9 @@ namespace Pims.Api.Services
             foreach (var deletedProperty in differenceSet)
             {
                 var acqFileProperties = _acquisitionFilePropertyRepository.GetPropertiesByAcquisitionFileId(acquisitionFile.Internal_Id).FirstOrDefault(ap => ap.PropertyId == deletedProperty.PropertyId);
-                if(acqFileProperties.PimsActInstPropAcqFiles.Any())
+                if(acqFileProperties.PimsActInstPropAcqFiles.Any() || acqFileProperties.PimsTakes.Any())
                 {
-                    throw new InvalidOperationException();
+                    throw new BusinessRuleViolationException();
                 }
                 _acquisitionFilePropertyRepository.Delete(deletedProperty);
                 if (deletedProperty.Property.IsPropertyOfInterest.HasValue && deletedProperty.Property.IsPropertyOfInterest.Value)
